\name{Asaphidae}
\docType{data}
\alias{Asaphidae}
\title{Phylogeny and age data for the Asaphidae}
\description{Phylogeny (162 most parsimonious trees) and age data for Asaphidae genera (Trilobita, Asaphida) taken from Bell and Braddy (2012).}
\usage{data(Asaphidae)}
\format{A list containing 162 trees (\code{$tree}) and a matrix of first and last appearances (\code{$ages}).}
\references{Bell, M. A. and Braddy, S. J., 2012, Cope's rule in the Ordovician trilobite Family Asaphidae (Order Asaphida): patterns across multiple most parsimonious trees: Historical Biology, v. 24, no. 3, p. 223-230.}
\keyword{datasets}