\name{Dipnoi}
\docType{data}
\alias{Dipnoi}
\title{Phylogeny and age data for dipnoans (lungfish)}
\description{Phylogeny (first most parsimonious tree) and age data for lungfish (Osteichthyes, Sarcopterygii, Dipnoi) taken from Lloyd et al. (2012).}
\usage{data(Dipnoi)}
\format{A list containing a tree (\code{$tree}) and a matrix of first and last appearances  (\code{$ages}).}
\references{Lloyd, G. T., Wang, S. C. and Brusatte, S. L., 2012. Identifying heterogeneity in rates of morphological evolution: discrete character change in the evolution of lungfish (Sarcopterygii; Dipnoi). Evolution, 66, 330-348.}
\keyword{datasets}