% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDescendants.R
\name{FindDescendants}
\alias{FindDescendants}
\title{Finds the tip numbers descending from a specific node in a phylo object}
\usage{
FindDescendants(n, tree)
}
\arguments{
\item{n}{The node number.}

\item{tree}{Tree as a phylo object.}
}
\value{
A vector of the descendant tip numbers.
}
\description{
Finds the tip numbers descending from a specific node in a phylo object.
}
\details{
A simple way to get the tips descending from a given node in a phylogenetic tree.
}
\examples{

# Find descendants of the root node in the lungfish tree:
FindDescendants(n = 87, tree = Dipnoi$tree)

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
