\name{UKzones}
\docType{data}
\alias{UKzones}
\title{British regional stages for the Ordovician}
\description{The stratigraphic ranges for the British stages of the Ordovician.}
\usage{data(UKzones)}
\format{A matrix containing the start and end ages for the British stage subdivisions of the Ordovician.}
\references{Webby, B.D., Paris, F., Droser, M.L., and Percival, I.G. (editors), 2004.  The Great Ordovician Biodiversity Event.  Columbia University Press, New York, 496 pp.}
\keyword{datasets}