% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strap-package.R
\docType{data}
\name{Asaphidae}
\alias{Asaphidae}
\title{Phylogeny and age data for the Asaphidae}
\format{A list containing 162 trees (\code{$tree}) and a matrix of first and last appearances (\link{$ages}).}
\description{
Phylogeny (162 most parsimonious trees) and age data for Asaphidae genera (Trilobita, Asaphida) taken from Bell and Braddy (2012).
}
\references{
Bell, M. A. and Braddy, S. J., 2012, Cope’s rule in the Ordovician trilobite Family Asaphidae (Order Asaphida): patterns across multiple most parsimonious trees: \emph{Historical Biology}, \bold{24}, 223-230.
}
\keyword{datasets}
