% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strap-package.R
\docType{data}
\name{Dipnoi}
\alias{Dipnoi}
\title{Phylogeny and age data for dipnoans (lungfish)}
\format{A list containing a tree (\code{$tree}) and a matrix of first and last appearances (\code{$ages}).}
\description{
Phylogeny (first most parsimonious tree) and age data for lungfish (Osteichthyes, Sarcopterygii, Dipnoi) taken from Lloyd et al. (2012).
}
\references{
Lloyd, G. T., Wang, S. C. and Brusatte, S. L., 2012. Identifying heterogeneity in rates of morphological evolution: discrete character change in the evolution of lungfish (Sarcopterygii; Dipnoi). \emph{Evolution}, \bold{66}, 330-348.
}
\keyword{datasets}
