% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strap-package.R
\docType{data}
\name{UKzones}
\alias{UKzones}
\title{British regional stages for the Ordovician}
\format{
A matrix containing the start and end ages for the British stage subdivisions of the Ordovician.
}
\description{
The stratigraphic ranges for the British stages of the Ordovician.
}
\references{
Webby, B.D., Paris, F., Droser, M.L., and Percival, I.G. (editors), 2004. \emph{The Great Ordovician Biodiversity Event}. Columbia University Press, New York, 496 pp.
}
\keyword{datasets}
